app/requirements.txt
celery==5.4.0
fastapi==0.115.11
mlflow==2.20.4
openpyxl==3.1.5
pika==1.3.2
Pillow==11.1.0
pydantic_settings==2.8.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
boto3==1.34.0
botocore==1.34.0
tensorflow==2.18.0